refactor: table naming of files
