+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/module/table.ts
@@ -2,11 +2,11 @@ import { Node, mergeAttributes } from '@tiptap/core'
 import { Table, TableCell, TableRow } from '@tiptap/extension-table'
 import IconTable from '@tabler/icons/outline/table.svg?raw'
 import { TiptapModule } from '../types.ts'
-import { tableCleanHtmlTransform } from '../table/table.ts'
-import { TableCaption, tableCaptionHtmlTransform, TableFigure } from '../table/tableCaption.ts'
-import { tableContextMenu } from '../table/tableContextMenu.ts'
-import { openTableDialog } from '../table/tableDialog.ts'
-import { CustomTableHeader, tableTheadHtmlTransform } from '../table/tableHeader.ts'
+import { tableCleanHtmlTransform } from '../table/cleanTransformer.ts'
+import { Caption, tableCaptionHtmlTransform, TableFigure } from '../table/caption.ts'
+import { contextMenu } from '../table/contextMenu.ts'
+import { openTableDialog } from '../table/dialogTable.ts'
+import { CustomTableHeader, tableTheadHtmlTransform } from '../table/header.ts'
 
 export const tableModule: TiptapModule = {
     extensions: getExtensions(),
@@ -22,7 +22,7 @@ export const tableModule: TiptapModule = {
         }
     ],
     contextMenuNode: 'table',
-    contextMenu: tableContextMenu
+    contextMenu: contextMenu
 }
 
 function getExtensions(): Node[] {
@@ -85,6 +85,6 @@ function getExtensions(): Node[] {
         CustomTableCell,
         CustomTableHeader,
         TableFigure,
-        TableCaption
+        Caption
     ]
 }
